Set the raft_multiplier flag to 1 on consul's > 0.7
Since consul 0.7, the consul config has a new flag `raft_multiplier'
which allow users to scale some raft timeout values. The default value
of the new flag is `5'. An example of such timeout is the leader
election timeout, which used to be 1s and now is 5x, i.e. `5s'. This
caused some test suites to timeout and fail since the cluster initially
takes 5s to elect a leader and be usable.

[#131641205]